+++ index.html
@@ -21,6 +21,6 @@
         a
     </footer>
     
-    <script src="script.js"></script>
+    <script src="script.js" type="module"></script>
 </body>
 </html>
